refactor: extracted add unit modal into its own component
